Added random char to leading and trailing password
In the vmware passgen.py.tmpl it is now possible to pass in a `leading_char` and a `trailing_char` and if you dont a random lowercase letter is picked to be used in ita place. This change forces either the passed in chars or a random lower case for all passwords now.

Signed-off-by: Michael Rice <michael@michaelrice.org>