rename vmware content patch bits
In the vmware content the esxi-patch-start and esxi-patch-complete tasks were poorly named. This change renamed those tasks to make it more obvious which part of the tasks put the host into and out of maint mode. This change should fix an issue that left the admin with no obvious reusable task to leave a host in maint mode.

Signed-off-by: Michael Rice <michael@michaelrice.org>