Fix up all the vmware boot things.
So it turns out that a leading / in the prefix= line of a boot.cfg
makes it much simpler to keep eferything in line.

This change makes legacy BIOS ise ipxe to load vmware instead of
esxi.i.  UEFI without a secure boot entitlement also continues to use
ipxe.  With secure boot entitlements, we use the signed bootloader embedded
in the ESXi install environment.