feat(vmware): added bootenvs for legacy
Commit that has the changes to the make-esxi.sh and all the bootenvs and templates.
In this change set the on disk name of RackN provided ISOs will now contain rkn_ as
a prefix. This change will require plugin users to delete their on disk extracted ISOs
and let the system re-extract them.

Signed-off-by: Michael Rice <michael@michaelrice.org>