feat(vmware): added script to remove drpy
Adds a script to the embedded files that will remove the drpy agent and firewwall vibs.
The script makes an attempt to cleanup the mess as well by removing config files stored on vfat and vmfs volumes.
Some files it may leave behind in /opt/rackn will disapear after the system is rebooted.

Signed-off-by: Michael Rice <michael@michaelrice.org>